# docs/swagger.yaml
definitions:
...
  model.AppAuthBindingItem:
    properties:
      id:
        type: string
      identifier:
        type: string
      identity_type:
        type: string
      is_primary:
        type: boolean
      status:
        type: string
      verified_at:
        description: RFC3339 string
        type: string
    type: object
  model.AppAuthBindingsResp:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AppAuthBindingItem'
        type: array
      user_id:
        type: string
    type: object
  model.AppBindEmailReq:
    properties:
      email:
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - email
    - verify_code
    type: object
  model.AppBindPhoneReq:
    properties:
      phone_number:
        maxLength: 50
        minLength: 5
        type: string
      phone_prefix:
        maxLength: 10
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - phone_number
    - verify_code
    type: object
  model.AppEmailLoginByCodeReq:
    properties:
      email:
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - email
    - verify_code
    type: object
  model.AppEmailRegisterReq:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - email
    - verify_code
    type: object
  model.AppEmailResetPasswordReq:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - email
    - password
    - verify_code
    type: object
  model.AppPhoneLoginByCodeReq:
    properties:
      phone_number:
        maxLength: 50
        minLength: 5
        type: string
      phone_prefix:
        maxLength: 10
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - phone_number
    - verify_code
    type: object
  model.AppPhoneRegisterReq:
    properties:
      password:
        maxLength: 255
        minLength: 6
        type: string
      phone_number:
        maxLength: 50
        minLength: 5
        type: string
      phone_prefix:
        maxLength: 10
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - phone_number
    - verify_code
    type: object
  model.AppPhoneResetPasswordReq:
    properties:
      password:
        maxLength: 255
        minLength: 6
        type: string
      phone_number:
        maxLength: 50
        minLength: 5
        type: string
      phone_prefix:
        maxLength: 10
        type: string
      verify_code:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - password
    - phone_number
    - verify_code
    type: object
  model.AppSendEmailCodeReq:
    properties:
      email:
        type: string
      scene:
        description: LOGIN/REGISTER/RESET_PASSWORD/BIND
        maxLength: 32
        type: string
    required:
    - email
    - scene
    type: object
  model.AppSendPhoneCodeReq:
    properties:
      phone_number:
        description: "13100000000"
        maxLength: 50
        minLength: 5
        type: string
      phone_prefix:
        description: "+86"
        maxLength: 10
        type: string
      scene:
        description: LOGIN/REGISTER/RESET_PASSWORD/BIND
        maxLength: 32
        type: string
    required:
    - phone_number
    - scene
    type: object
  model.AppUnbindReq:
    properties:
      identity_type:
        enum:
        - PHONE
        - EMAIL
        maxLength: 32
        type: string
    required:
    - identity_type
    type: object
  model.AppWxmpBindPhoneReq:
    properties:
      phone_code:
        description: wx.getPhoneNumber 返回的 code
        type: string
    required:
    - phone_code
    type: object
  model.AppWxmpLoginReq:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  model.AppWxmpProfileReq:
    properties:
      avatar_url:
        maxLength: 500
        type: string
      city:
        maxLength: 50
        type: string
      country:
        maxLength: 50
        type: string
      gender:
        description: 0未知 1男 2女
        enum:
        - 0
        - 1
        - 2
        type: integer
      language:
        maxLength: 20
        type: string
      nick_name:
        maxLength: 50
        type: string
      province:
        maxLength: 50
        type: string
    type: object
...
  model.LoginRsp:
    properties:
      expires_in:
        description: 过期时间(单位:秒)
        type: integer
      token:
        description: 登录凭证
        type: string
    type: object
...
  model.Response:
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 业务数据
      message:
        description: 提示信息
        example: success
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: ThingsPanel API.
  title: ThingsPanel API
  version: "1.0"
paths:
...
  /api/v1/app/auth/bind/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppBindEmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 绑定邮箱
      tags:
      - APP-Auth
  /api/v1/app/auth/bind/phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppBindPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 绑定手机号
      tags:
      - APP-Auth
  /api/v1/app/auth/bindings:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AppAuthBindingsResp'
      summary: 当前账号绑定信息
      tags:
      - APP-Auth
...
  /api/v1/app/auth/email/code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppSendEmailCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: APP发送邮箱验证码
      tags:
      - APP-Auth
  /api/v1/app/auth/email/login_by_code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppEmailLoginByCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginRsp'
      summary: 邮箱验证码登录
      tags:
      - APP-Auth
  /api/v1/app/auth/email/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppEmailRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginRsp'
      summary: 邮箱注册
      tags:
      - APP-Auth
  /api/v1/app/auth/email/reset_password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppEmailResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 邮箱重置密码
      tags:
      - APP-Auth
  /api/v1/app/auth/phone/code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppSendPhoneCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: APP发送手机号验证码
      tags:
      - APP-Auth
  /api/v1/app/auth/phone/login_by_code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppPhoneLoginByCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginRsp'
      summary: 手机号验证码登录
      tags:
      - APP-Auth
  /api/v1/app/auth/phone/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppPhoneRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginRsp'
      summary: 手机号注册
      tags:
      - APP-Auth
  /api/v1/app/auth/phone/reset_password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppPhoneResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 手机号重置密码
      tags:
      - APP-Auth
  /api/v1/app/auth/unbind:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppUnbindReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 解绑账号（手机/邮箱）
      tags:
      - APP-Auth
  /api/v1/app/auth/wxmp/bind_phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppWxmpBindPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 微信小程序一键绑定手机号
      tags:
      - APP-Auth
  /api/v1/app/auth/wxmp/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppWxmpLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginRsp'
      summary: 微信小程序一键登录
      tags:
      - APP-Auth
  /api/v1/app/auth/wxmp/profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-TenantID
        required: true
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AppWxmpProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 微信小程序用户信息解析/保存
      tags:
      - APP-Auth
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"